requirements.txt
fastapi==0.109.2
pydantic==2.6.1
PyPDF2==3.0.1
langchain==0.1.6
langchain-community==0.0.19
python-dotenv==1.0.1
langchain-openai==0.0.5
uvicorn==0.27.0.post1
faiss-cpu==1.7.4
SQLAlchemy==2.0.25